1. 調整 SearchBar 架構 2. 新增顯示商品數量選單
